PAE-1382: Extract balance effects from update-status.js
Moves deductWasteBalanceIfNeeded, deductTotalBalanceIfNeeded,
creditWasteBalanceIfNeeded, creditFullBalanceIfNeeded, and
applyWasteBalanceEffects into a dedicated module. Pure refactor
with no behaviour change.
